Initial commit of game.
A Game owns a Deck, which owns Cards. A Game also has fuses
and clocks. The Game right now just deals out all the cards
in its deck until there are no more cards left.

Deck shuffling is broken. Game has a state for every turn,
which is just a concat of the bombs, deck, and clock.