Game is pretty much playable, I think.
It works as-is, but two people can't actually play it sharing a
terminal, because they'll see what the previous player sees, i.e.
their own cards. But the game logic I think is all there.

UI is bad and probably not very robust. Haven't done thorough bug
testing. Haven't tested with more than 2 players, and rainbow suit
is not ready.

Should at least move the consequences of player actions into
Game, if not the decision making itself.

Nearly ready to teach a computer to play it, though.